Move contains(T item) to ListDataView
Fixes #8654
