+++ flow-data/src/main/java/com/vaadin/flow/data/provider/DataView.java
@@ -19,9 +19,7 @@ package com.vaadin.flow.data.provider;
 import java.io.Serializable;
 import java.util.stream.Stream;
 
-import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.ComponentEventListener;
-import com.vaadin.flow.function.SerializableConsumer;
 import com.vaadin.flow.shared.Registration;
 
 /**
@@ -41,30 +39,6 @@ public interface DataView<T> extends Serializable {
      * @return filtered and sorted data set
      */
     Stream<T> getItems();
-
-    /**
-     * Check if item is in the current data. Item may be filtered out or for
-     * lazy data not in the currently loaded making it un-available.
-     * <p>
-     * By default, {@code equals} method implementation of the item is used for
-     * identity check. If a custom data provider is used, then the
-     * {@link DataProvider#getId(Object)} method is used instead. Item's custom
-     * identity can be set up with a
-     * {@link DataView#setIdentifierProvider(IdentifierProvider)}.
-     * <p>
-     * <em>NOTE:</em> when the component is created and not yet added, the item
-     * might not yet be loaded, but will be loaded during the "before client
-     * response"-phase. To check if the item has been added at that point, after
-     * setting the data source to the component you need to use
-     * {@link com.vaadin.flow.component.UI#beforeClientResponse(Component, SerializableConsumer)}.
-     *
-     * @param item
-     *            item to search for
-     * @return true if item is found in the available data
-     *
-     * @see #setIdentifierProvider(IdentifierProvider)
-     */
-    boolean contains(T item);
 
     /**
      * Add an item count change listener that is fired when the item
