+++ flow-data/src/test/java/com/vaadin/flow/data/provider/AbstractDataViewTest.java
@@ -19,19 +19,17 @@ package com.vaadin.flow.data.provider;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Objects;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Stream;
-
-import com.vaadin.flow.tests.data.bean.Item;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.ComponentUtil;
 import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.function.SerializableSupplier;
+import com.vaadin.flow.tests.data.bean.Item;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
 
 public class AbstractDataViewTest {
 
@@ -45,7 +43,6 @@ public class AbstractDataViewTest {
 
     @Before
     public void init() {
-
         items = new ArrayList<>(Arrays.asList(
                 new Item(1L, "first", "description1"),
                 new Item(2L, "middle", "description2"),
@@ -91,71 +88,10 @@ public class AbstractDataViewTest {
         Assert.assertEquals(10, fired.get());
     }
 
-    @Test
-    public void setIdentifierProvider_defaultIdentity_equalsIsUsed() {
-        Assert.assertTrue(dataView.contains(
-                new Item(1L, "first")));
-        Assert.assertFalse(dataView.contains(
-                new Item(1L, "non present")));
-        Assert.assertFalse(dataView.contains(
-                new Item(4L, "first")));
-    }
-
-    @Test
-    public void setIdentifierProvider_dataProviderIdentity_getIdIsUsed() {
-        dataProvider = new CustomIdentityItemDataProvider(items);
-
-        Assert.assertTrue(dataView.contains(
-                new Item(1L, "first")));
-        Assert.assertTrue(dataView.contains(
-                new Item(1L, "non present")));
-        Assert.assertFalse(dataView.contains(
-                new Item(4L, "first")));
-    }
-
-    @Test
-    public void setIdentifierProvider_customIdentifierProvider_customIdentifierProviderIsUsed() {
-        dataView.setIdentifierProvider(Item::getValue);
-
-        Assert.assertTrue(dataView.contains(
-                new Item(1L, "first")));
-        Assert.assertFalse(dataView.contains(
-                new Item(1L, "non present")));
-        Assert.assertTrue(dataView.contains(
-                new Item(4L, "first")));
-    }
-
-    @Test
-    public void setIdentifierProvider_dataProviderHasChanged_newDataProviderIsUsed() {
-        Assert.assertFalse(dataView.contains(
-                new Item(1L, "non present")));
-
-        dataProvider = new CustomIdentityItemDataProvider(items);
-
-        Assert.assertTrue(dataView.contains(
-                new Item(1L, "non present")));
-
-        dataProvider = DataProvider.ofCollection(items);
-
-        Assert.assertFalse(dataView.contains(
-                new Item(1L, "non present")));
-    }
-
-    @Test
-    public void setIdentifierProvider_dataProviderHasChanged_identifierProviderRetained() {
-        Assert.assertFalse(dataView.contains(
-                new Item(4L, "non present", "description1")));
-
-        dataView.setIdentifierProvider(Item::getDescription);
-
-        Assert.assertTrue(dataView.contains(
-                new Item(4L, "non present", "description1")));
-
-        dataProvider = new CustomIdentityItemDataProvider(items);
-
-        Assert.assertTrue(dataView.contains(
-                new Item(4L, "non present", "description1")));
-    }
+    /**
+     * setIdentifierProvider is tested in AbstractListDataView since it
+     * has the container(T item) method.
+     */
 
     @Tag("test-component")
     private static class TestComponent extends Component {
@@ -172,15 +108,6 @@ public class AbstractDataViewTest {
         @Override
         protected Class<?> getSupportedDataProviderType() {
             return DataProvider.class;
-        }
-
-        @Override
-        public boolean contains(Item item) {
-            IdentifierProvider<Item> identifierProvider =
-                    getIdentifierProvider();
-            return getItems().anyMatch(i -> Objects.equals(
-                    identifierProvider.apply(item),
-                    identifierProvider.apply(i)));
         }
     }
